fix community component

+++ src/modules/communities/services/community.service.ts
@@ -1,23 +1,29 @@
-import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { Community } from '../schemas/community.schema';
 import { Model ,ObjectId} from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
-import { CommunityDto } from '../dto/newCommunity.dto';
+import { CommunityDto } from '../models/newCommunity.dto';
+import { ICommunity } from '../models/community.model';
+import { UserService } from 'src/modules/users/services/user.service';
 
 @Injectable()
 export class CommunityService {
-    constructor(@InjectModel('Community') private readonly communityModel: Model<Community>) {}
+    constructor(@InjectModel('Community')
+        private readonly communityModel: Model<Community>,
+        private readonly userService: UserService
+    ) { }
     
-    async create(requestData: CommunityDto): Promise<Community> {
-    const existingCommunity = await this.communityModel.findOne({ name: requestData.name });
+    async create(requestData: CommunityDto) {
+        const existingCommunity = await this.communityModel.findOne({ name: requestData.name });
 
-    if (existingCommunity) {
-        throw new BadRequestException('');
+        if (existingCommunity) {
+            throw new BadRequestException('');
+        }
+
+        const newCommunity = await this.communityModel.create(requestData);
+        const community = newCommunity.toObject() as unknown as ICommunity
+        return await this.userService.addCommunity(requestData.creator,community._id);
     }
-
-    const newCommunity = new this.communityModel(requestData);
-    return newCommunity.save();
-}
 
     async findOne(communityName: string): Promise<Community >{
         const community = this.communityModel.findOne({name:communityName});
@@ -35,31 +41,22 @@ export class CommunityService {
         return community 
     }
 
-    async addMods(communityId: ObjectId, mods: ObjectId[]) {
+    async addMods(communityId: ObjectId, userId: ObjectId, mods: ObjectId[]) {
         const community = await this.communityModel.findById(communityId);
 
         if (!community) {
             throw new NotFoundException('');
+        }
+        if (!this.validateCreator(community, userId)) {
+            throw new UnauthorizedException('');
         }
         const newMods = mods.filter(mod => !community.mods.includes(mod));
         community.mods = [...community.mods, ...newMods];
-        community.save();
+        await community.save();
         return ;
     }
 
-    async addTags(communityId: ObjectId, tags: string[]) {
-        const community = await this.communityModel.findById(communityId);
-
-        if (!community) {
-            throw new NotFoundException('');
-        }
-        const newTags = tags.filter(tag => !community.tags.includes(tag));
-        community.tags = [...community.tags, ...newTags];
-        community.save();
-        return ;
-    }
-
-    async removeMod(communityId: ObjectId, modId: ObjectId) {
+    async removeMod(communityId: ObjectId, userId: ObjectId, modId: ObjectId) {
         const community = await this.communityModel.findById(communityId);
 
         if (!community) {
@@ -68,34 +65,65 @@ export class CommunityService {
         if (!community.mods.includes(modId)) {
             throw new NotFoundException('');
         }
-
+        if (!this.validateCreator(community, userId)) {
+            throw new UnauthorizedException('');
+        }
         const modIdString = modId.toString();
 
         community.mods = community.mods.filter(mod => mod.toString() !== modIdString);
-        community.save();
+        await community.save();
         return ;
     }
 
-    async removeTag(communityId: ObjectId, tagString: string) {
+    async addTags(communityId: ObjectId, userId: ObjectId, tags: string[]) {
+        const community = await this.communityModel.findById(communityId);
+
+        if (!community) {
+            throw new NotFoundException('');
+        }
+        if (!this.validateCreator(community, userId) && !this.validateMod(community, userId)) {
+            throw new UnauthorizedException('');
+        }
+        const newTags = tags.filter(tag => !community.tags.includes(tag));
+        community.tags = [...community.tags, ...newTags];
+        await community.save();
+        return ;
+    }
+
+    async removeTag(communityId: ObjectId,  userId: ObjectId, tagString: string) {
         const community = await this.communityModel.findById(communityId);
 
         if (!community) {
             throw new NotFoundException();
+        }
+        if (!this.validateCreator(community, userId) && !this.validateMod(community, userId)) {
+            throw new UnauthorizedException('');
         }
         if (!community.tags.includes(tagString)) {
             throw new NotFoundException();
         }
 
         community.tags = community.tags.filter(tag => tag!== tagString);
-        community.save();
+        await community.save();
         return ;
     }
 
-    async delete(id: ObjectId) {
+    async delete(id: ObjectId , userId: ObjectId) {
         const community = await this.communityModel.findById(id);
         if (!community) {
             throw new NotFoundException('');
         }
+        if (!this.validateCreator(community, userId)) {
+            throw new UnauthorizedException('');
+        }
+        await this.userService.removeCommunity(userId, id);
         return await this.communityModel.findByIdAndDelete(id);
+    }
+
+    private validateCreator(community: Community, userId: ObjectId): boolean {
+        return community.creator.toString() == userId.toString()
+    }
+    private validateMod(community: Community, modId: ObjectId): Boolean{
+        return community.mods.includes(modId);
     }
 }
